refactored complete for RhapsodyCreatorGenerative.test
