+++ test/core/testContractSale.ts
@@ -2,13 +2,13 @@ const { expect } = require('chai');
 import { ethers } from 'ethers';
 import { defaultAbiCoder, keccak256 } from 'ethers/lib/utils';
 import { beforeEachSetupForGenerative } from '../helpers/contractBeforeEachSetup';
-import { params } from '../RhapsodyCreatorGenerative.test';
+import { currentBlockTime, params } from '../RhapsodyCreatorGenerative.test';
 import { keccak256Hashes } from '../helpers/generateKeccak256Hash';
-import { Merklized, RhapsodyCreatorSaleType } from './type';
+import { Merklized, RhapsodyCreatorSaleType, RhapsodyCreatorVariation } from './type';
 const { generateLeaf, buildWhitelist } = require('../../scripts/helpers/whitelist');
 const { parseEther } = require('../helpers/constant');
 
-export const testContractSale = (types: RhapsodyCreatorSaleType[]) => {
+export const testContractSale = (variation: RhapsodyCreatorVariation, types: RhapsodyCreatorSaleType[]) => {
   describe('sale', () => {
     types.forEach((type) => {
       switch (type) {
@@ -24,6 +24,131 @@ export const testContractSale = (types: RhapsodyCreatorSaleType[]) => {
       }
     });
   });
+
+  describe('extra', () => {
+    switch (variation) {
+      case 'generative':
+        _testContractGenerative();
+        break;
+    }
+  });
+};
+
+const _testContractGenerative = () => {
+  describe('setMintTime', () => {
+    let creator: ethers.Contract;
+
+    let minter: (minter: any, invocations: any, proof: any) => void;
+    beforeEach(async () => {
+      const { contracts } = await beforeEachSetupForGenerative(params);
+
+      creator = contracts.creator;
+
+      minter = async (minter, invocations, proof) => creator.connect(minter).claimMint(invocations, proof);
+    });
+
+    it('should be able to set valid mint times', async () => {
+      await creator.setMintTime(currentBlockTime + 100, currentBlockTime + 105, currentBlockTime + 110);
+
+      expect(await creator.claimTime()).to.be.equal(currentBlockTime + 100);
+
+      expect(await creator.presaleTime()).to.be.equal(currentBlockTime + 105);
+
+      expect(await creator.publicTime()).to.be.equal(currentBlockTime + 110);
+    });
+
+    it('can change presale and public time retroactively', async () => {
+      await creator.setMintTime(currentBlockTime + 100, currentBlockTime + 105, currentBlockTime + 110);
+
+      // forward time
+      await creator.setMintTime(currentBlockTime + 105, currentBlockTime + 110, currentBlockTime + 115);
+
+      expect(await creator.claimTime()).to.be.equal(currentBlockTime + 105);
+
+      expect(await creator.presaleTime()).to.be.equal(currentBlockTime + 110);
+
+      expect(await creator.publicTime()).to.be.equal(currentBlockTime + 115);
+
+      // backward time
+      await creator.setMintTime(currentBlockTime + 100, currentBlockTime + 105, currentBlockTime + 110);
+
+      expect(await creator.claimTime()).to.be.equal(currentBlockTime + 100);
+      expect(await creator.presaleTime()).to.be.equal(currentBlockTime + 105);
+      expect(await creator.publicTime()).to.be.equal(currentBlockTime + 110);
+
+      // testing edge case
+      await creator.setMintTime(currentBlockTime + 1, currentBlockTime + 2, currentBlockTime + 3);
+      expect(await creator.claimTime()).to.be.equal(currentBlockTime + 1);
+      expect(await creator.presaleTime()).to.be.equal(currentBlockTime + 2);
+      expect(await creator.publicTime()).to.be.equal(currentBlockTime + 3);
+    });
+
+    it('public sale time cannot be less than presale time', async () => {
+      await expect(
+        creator.setMintTime(currentBlockTime + 100, currentBlockTime + 110, currentBlockTime + 105)
+      ).to.be.revertedWith('RhapsodyCreatorGenerative/invalid-public-time');
+
+      // edge cases
+      await expect(
+        creator.setMintTime(currentBlockTime + 1, currentBlockTime + 2, currentBlockTime + 1)
+      ).to.be.revertedWith('RhapsodyCreatorGenerative/invalid-public-time');
+
+      await expect(
+        creator.setMintTime(currentBlockTime + 1, currentBlockTime + 2, currentBlockTime + 0)
+      ).to.be.revertedWith('RhapsodyCreatorGenerative/invalid-public-time');
+    });
+
+    it('presale time cannot be less than the current block timestamp', async () => {
+      await expect(
+        creator.setMintTime(currentBlockTime + 1, currentBlockTime - 10, currentBlockTime + 10)
+      ).to.be.revertedWith('RhapsodyCreatorGenerative/invalid-presale-time');
+
+      await expect(
+        creator.setMintTime(currentBlockTime + 1, currentBlockTime - 1, currentBlockTime)
+      ).to.be.revertedWith('RhapsodyCreatorGenerative/invalid-presale-time');
+
+      await expect(
+        creator.setMintTime(currentBlockTime + 1, currentBlockTime - 2, currentBlockTime - 1)
+      ).to.be.revertedWith('RhapsodyCreatorGenerative/invalid-presale-time');
+    });
+  });
+
+  describe('maxCollection', function () {
+    this.timeout(10000);
+    let minterA: any;
+    let creator: ethers.Contract;
+    beforeEach(async () => {
+      const paramsB = {
+        collectionSize: 1200,
+        amountForPromotion: 0,
+        maxPublicBatchPerAddress: 1200,
+        mintPrice: parseEther(0.01),
+      };
+
+      const { contracts, wallets } = await beforeEachSetupForGenerative(paramsB);
+
+      creator = contracts.creator;
+      minterA = wallets.minterA;
+    });
+
+    it('should all be unique token hashes', async () => {
+      let hashes = {};
+      for (let i = 0; i < 12; i++) {
+        await creator.connect(minterA).publicMint(100, {
+          value: parseEther(0.01 * 100),
+        });
+        for (let id = 0; id < 100; id++) {
+          const hash = await creator.tokenHash(id * (i + 1));
+
+          // @ts-ignore
+          if (hashes[hash] !== undefined) {
+            console.log('found existing token hash');
+            throw new Error('it failed');
+          }
+        }
+      }
+    });
+  });
 };
 
 const _testContractClaim = () => {
@@ -31,7 +156,6 @@ const _testContractClaim = () => {
     let minterA: any, minterB: any, minterC: any;
 
     let creator: ethers.Contract;
-    let randomizer: ethers.Contract;
 
     let claimMerklized: Merklized;
     let presaleMerklized: Merklized;
@@ -45,7 +169,6 @@ const _testContractClaim = () => {
       minterC = wallets.minterC;
 
       creator = contracts.creator;
-      randomizer = contracts.randomizer;
 
       claimMerklized = merkle.claimMerklized;
       presaleMerklized = merkle.presaleMerklized;
